% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imgProc.R
\name{blobDetector}
\alias{blobDetector}
\title{Find blobs in binary image}
\usage{
blobDetector(image)
}
\arguments{
\item{image}{A binary vpImage object, for instance one obtained from the
\code{\link{thresholding}} function.}
}
\value{
This function returns a data frame with one row per blob detected in
the image and seven columns corresponding to the seven following properties
of each blob:
\itemize{
 \item{"id": identity of the blob.}
 \item{"x": x coordinate of the center of the ellipse approximating the shape
 of the blob.}
 \item{"y": y coordinate of the center of the ellipse approximating the shape
 of the blob.}
 \item{"alpha": orientation in degrees of the major axis of the ellipse.}
 \item{"major": length of the major axis of the ellipse.}
 \item{"minor": length of the minor axis of the ellipse.}
 \item{"area": area of the polygon containing the blob.}
}
}
\description{
This function finds blobs in binary images. A blob is defined as a group of
8-connected white pixels on a black background.
}
\examples{
require(plotrix)
filename <- system.file("sample_img/dots.jpg", package = "videoplayR")
img <- readImg(filename)
bw <- thresholding(img, 200, "inverted")
blobs <- blobDetector(bw)

imshow(img)
points(y ~ x, data = blob, col = "red", pch = 20)
draw.ellipse(blob$x, blob$y, blob$major / 2, blob$minor / 2, blob$alpha, border = "red")
}
\seealso{
\code{\link{thresholding}}
}

