% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imgProc.R
\name{ddd2d}
\alias{ddd2d}
\title{Flatten a 3-channels image to 1 channel}
\usage{
ddd2d(image)
}
\arguments{
\item{image}{A 3-channels vpImage object to convert to 1 channel.}
}
\value{
This function returns a vpImage object.
}
\description{
This function converts a 3-channels image (rgb or numeric) to a 1-channel
image (binary, grayscale or numeric) by computing the average of the 3
channel values for each pixel. For instance, this function can be used to
convert an rgb image to grayscale.
}
\details{
The vpImage object created by the \code{ddd2d} function is not a
persistent object, but a pointer to a C++ object that cannot be saved for
reuse in a different session.
}
\examples{
filename <- system.file("sample_img/SampleVideo_1080x720_5mb.png", package = "videoplayR")
img <- readImg(filename)
imshow(img)
imshow(ddd2d(img))
}
\seealso{
\code{\link{d2ddd}}, \code{\link{vpImage}}
}

