% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{getFrame}
\alias{getFrame}
\title{Grab a specific frame from a video}
\usage{
getFrame(video, frame)
}
\arguments{
\item{video}{The vpVideo object from which the frame should be grabbed.}

\item{frame}{The position in the video of the frame to be grabbed.}
}
\value{
This function returns a vpImage object.
}
\description{
This function grabs a specific frame from a vpVideo object and returns a 
vpImage object that can be manipulated using the \code{\link{videoplayR}} 
image and video manipulation toolbox.
}
\details{
The vpImage object created by the \code{grabFrame} function is not a 
persistent object, but a pointer to a C++ object that cannot be saved for 
reuse in a different session.
}
\examples{
filename <- system.file("sample_vid/SampleVideo_1080x720_5mb.mp4", package = "videoplayR")
vid <- readVid(filename)
frame <- getFrame(vid, 1)
imshow(frame)
}
\seealso{
\code{\link{readVid}}, \code{\link{vpVideo}}, \code{\link{vpImage}}
}

