% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{img2r}
\alias{img2r}
\title{Convert a vpImage object to a matrix or an array}
\usage{
img2r(image)
}
\arguments{
\item{image}{The vpImage object to convert.}
}
\value{
This function returns either a 2D matrix or a 3D arrays depending on
the input vpImage object.
}
\description{
This function converts a vpImage object to a 2D matrix or a 3D array that can 
be used in R. Binary, grayscale, and 2D numeric vpImage objects will be 
converted to 2D matrices. RGB and 3D numeric vpImage objects will be 
converted to 3D arrays.
}
\examples{
filename <- system.file("sample_img/SampleVideo_1080x720_5mb.png", package = "videoplayR")
img <- readImg(filename)
imgMat <- img2r(img)
imgMat
}
\seealso{
\code{\link{r2img}}, \code{\link{vpImage}}
}

