% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imshow.R
\name{imshow}
\alias{imshow}
\title{Display vpImage object}
\usage{
imshow(image)
}
\arguments{
\item{image}{A vpImage object to display.}
}
\value{
This function does not return anything.
}
\description{
This function opens a graphics device and display the image contained in a
vpImage object.
}
\details{
\code{imshow} opens a regular graphics device, meaning that it is
possible to overlay lines and points over the image, like with any regular
plot. The bottom left corner of the image is set at \{1,1\}.

If the type of the vpImage object is "numeric", then its values are rescaled
between 0 and 1, with 0 representing the minimum value of the original image
and 1 representing its maximum.
}
\examples{
filename <- system.file("sample_img/SampleVideo_1080x720_5mb.png", package = "videoplayR")
img <- readImg(filename)
imshow(img)
}
\seealso{
\code{\link{vpImage}}
}

