% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{nextFrame}
\alias{nextFrame}
\title{Grab the next frame from a video stream}
\usage{
nextFrame(stream)
}
\arguments{
\item{stream}{The vpStream object from which the frame should be grabbed.}
}
\value{
This function returns a vpImage object.
}
\description{
This function grabs the next available frame from a vpStream object and 
returns a vpImage object that can be manipulated using the 
\code{\link{videoplayR}} image and video manipulation toolbox.
}
\details{
The vpImage object created by the \code{grabFrame} function is not a 
persistent object, but a pointer to a C++ object that cannot be saved for 
reuse in a different session.
}
\examples{
stream <- readStream(0)
frame <- nextFrame(stream)
imshow(frame)
release(stream)
}
\seealso{
\code{\link{readStream}}, \code{\link{release}}, \code{\link{vpStream}}, 
\code{\link{vpImage}}
}

