% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{r2img}
\alias{r2img}
\title{Convert a matrix or an array to a vpImage object}
\usage{
r2img(array, numeric = FALSE)
}
\arguments{
\item{array}{The matrix or array  to convert.}

\item{numeric}{A logical value indicating if the matrix or array should be
treated as a numeric object. If FALSE (default), the matrix or array will be
casted to an integer matrix or array. If TRUE, the matrix or array will be
converted to a numeric vpImage object.}
}
\value{
This function returns a vpImage object.
}
\description{
This function converts a 2D matrix or a 3D array to a vpImage object to that
can be manipulated using the \code{\link{videoplayR}} image and video
manipulation toolbox. Binary matrices will be converted to binary vpImage
objects. Integer matrices with values in [0, 255] will be converted to
grayscale vpImage objects. Numeric matrics will be converted to 2D numeric
vpImage objects. 3D integer arrays with values in [0, 255] will be converted
to RGB vpImage objects. 3D numeri arrays will be converted to 3D numeric
vpImage objects.
}
\examples{
gradient <- function(x1, x2) { x1 + x2 }
x <- 0:127
mat <- gradient(outer(rep(1, length(x)), x), x)
img <- r2img(mat)
imshow(img)
}
\seealso{
\code{\link{img2r}}, \code{\link{vpImage}}
}

