% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readImg}
\alias{readImg}
\title{Load image file in memory}
\usage{
readImg(filename)
}
\arguments{
\item{filename}{A character string of the path to the image file.}
}
\value{
This function returns a vpImage object.
}
\description{
This function reads an image file and return a vpImage object that can be 
manipulated using the \code{\link{videoplayR}} image and video manipulation 
toolbox. This function supports the following image formats: 
\itemize{
 \item{Windows bitmaps - *.bmp, *.dib}
 \item{JPEG files - *.jpeg, *.jpg, *.jpe}
 \item{Portable Network Graphics - *.png}
 \item{Portable image format - *.pbm, *.pgm, *.ppm}
 \item{TIFF files - *.tiff, *.tif}
}
}
\details{
The vpImage object created by the \code{readImg} function is not a 
persistent object, but a pointer to a C++ object that cannot be saved for 
reuse in a different session.
}
\examples{
filename <- system.file("sample_img/SampleVideo_1080x720_5mb.png", package = "videoplayR")
img <- readImg(filename)
imshow(img)
}
\seealso{
\code{\link{writeImg}}, \code{\link{vpImage}}
}

