% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readStream}
\alias{readStream}
\title{Capture video stream from camera}
\usage{
readStream(cam = 0L)
}
\arguments{
\item{cam}{The camera id number (default: 0).}
}
\value{
This function returns a vpVideo object.
}
\description{
This function captures a video stream from a camera and return a vpVideo 
object that can be manipulated using the \code{\link{videoplayR}} image and 
video manipulation toolbox. This function should support most webcams and 
other cameras supported by \href{http://opencv.org/}{OpenCV}.
}
\details{
The vpStream object created by the \code{readStream} function is not 
a persistent object, but a pointer to a C++ object that cannot be saved for 
reuse in a different session.
}
\examples{
stream <- readStream(0)
frame <- nextFrame(stream)
imshow(frame)
release(stream)
}
\seealso{
\code{\link{release}}, \code{\link{readImg}}, \code{\link{readVideo}}, 
\code{\link{nextFrame}}, \code{\link{vpStream}}
}

