% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\name{readVid}
\alias{readVid}
\title{Load video file in memory}
\usage{
readVid(filename)
}
\arguments{
\item{filename}{A character string of the path to the video file.}
}
\value{
This function returns a vpVideo object.
}
\description{
This function reads a video file and return a vpVideo object that can be
manipulated using the \code{\link{videoplayR}} image and video manipulation
toolbox. This function should support most video formats and codecs listed at
\link{http://www.fourcc.org/codecs.php}.
}
\details{
The vpVideo object created by the \code{readVid} function is not a
persistent object, but a pointer to a C++ object that cannot be saved for
reuse in a different session.
}
\examples{
filename <- system.file("sample_vid/SampleVideo_1080x720_5mb.mp4", package = "videoplayR")
vid <- readVid(filename)
frame <- getFrame(vid, 1)
imshow(frame)
}
\seealso{
\code{\link{readImg}}, \code{\link{readStream}}, \code{\link{getFrame}},
\code{\link{vpVideo}}
}

