% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{release}
\alias{release}
\title{Release a video stream}
\usage{
release(stream)
}
\arguments{
\item{stream}{The vpVideo or vpStream object to release.}
}
\value{
This function does not return anything.
}
\description{
This function releases (i.e. closes) the video stream contained in a vpVideo 
or a vpStream object.
}
\details{
Only the video stream contained in the vpVideo or vpStream object is
released (i.e closed). The vpVideo or vpStream object will remain in the 
session environment until removed by the user. Once released, the video 
stream cannot be accessed anymore unless a new vpVideo or vpStream object is 
created.
}
\examples{
stream <- readStream(0)
frame <- nextFrame(stream)
imshow(frame)
release(stream)
}
\seealso{
\code{\link{readStream}}, \code{\link{vpStream}}, \code{\link{vpVideo}}
}

