% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgProc.R
\name{thresholding}
\alias{thresholding}
\title{Convert image to binary image}
\usage{
thresholding(image, thres, type)
}
\arguments{
\item{image}{A vpImage object to convert to a binary image.}

\item{thres}{The threshold value to separate the pixels of the image in two 
groups.}

\item{type}{The type of thresholding function to apply to the image. If set 
to \code{"binary"}, then the pixels above \code{thres} are replaced with ones
and those below are replaced with zeros. If set to \code{"inverted"}, then 
the pixels above \code{thres} are replaced with zeros and those below are 
replaced with ones.}
}
\value{
This function returns a vpImage object.
}
\description{
This function converts an image (rgb, grayscale or numeric) to binary by 
applying a thresholding function to it. 3-channels images (rgb and numeric)
are first converted to 1-channel images (grayscale or numeric) prior to 
applying the thresholding function.
}
\details{
The vpImage object created by the \code{thresholding} function is 
not a persistent object, but a pointer to a C++ object that cannot be saved 
for reuse in a different session.
}
\examples{
filename <- system.file("sample_img/SampleVideo_1080x720_5mb.png", package = "videoplayR")
img <- readImg(filename)
imshow(img)
bin <- thresholding(img, 128, "binary")
imshow(bin)
inv <- thresholding(img, 128, "inverted")
imshow(inv)
}
\seealso{
\code{\link{vpImage}}
}

