\name{videoplayR-package}
\alias{videoplayR-package}
\alias{videoplayR}
\docType{package}
\title{
Functions to manipulate video files with R
}
\description{
This package uses the OpenCV library to provide R users with some 
basic functions to read and manipulate video files. 
}
\details{
\tabular{ll}{
Package: \tab videoplayR\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2014-12-23\cr
License: \tab GPL-3\cr
}
}
\author{
Simon Garnier <garnier@njit.edu> - \href{http://twitter.com/sjmgarnier}{@sjmgarnier}

Maintainer: Simon Garnier <garnier@njit.edu> - \href{http://twitter.com/sjmgarnier}{@sjmgarnier}

}
\references{
More about OpenCV at http://opencv.org/
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%% ~~ simple examples of the most important functions ~~
%}
