% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videoplayR-package.R
\docType{package}
\name{videoplayR}
\alias{videoplayR}
\alias{videoplayR-package}
\title{videoplayR: A basic (for now) computer vision library for R.}
\description{
The \code{videoplayR} package uses the OpenCV library to provide R users with 
some basic functions to read and manipulate image and video files.
}
\details{
\tabular{ll}{
 Package: \tab videoplayR\cr
 Type: \tab Package\cr
 Version: \tab 0.3.1\cr
 Date: \tab 2015-09-30\cr
 License: \tab GPL-3\cr
}
}
\author{
Simon Garnier <garnier@njit.edu> - \href{http://twitter.com/sjmgarnier}{@sjmgarnier}

Maintainer: Simon Garnier <garnier@njit.edu> - \href{http://twitter.com/sjmgarnier}{@sjmgarnier}
}
\references{
More about OpenCV at \link{http://opencv.org/}.
}

