% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\docType{data}
\name{vpImage}
\alias{vpImage}
\title{Image object.}
\format{A vpImage object contains two fields:
\itemize{
 \item{\code{$type}: the type of image stored in the object. There are 4 types of image types:
   \itemize{
     \item{\code{binary}: 1-channel image with integer pixel values of 0 or 1 only.}
     \item{\code{grayscale}: 1-channel image with integer pixel values in [0, 255].}
     \item{\code{rgb}: 3-channels image with integer pixel values in [0, 255].}
     \item{\code{numeric}: 1- or 3-channels image with unconstrained pixel values.}
   }
 }
 \item{"dim": the dimensions in pixels (width x height x depth) of the image.}
}}
\description{
A vpImage object is an R container for an OpenCV image. vpImage objects can
be generated by the following \code{\link{videoplayR}} functions:
\code{\link{readImg}}, \code{\link{getFrame}}. They can be displayed using
the \code{\link{inshow}} function.
}
\details{
A vpImage object is not a persistent object, but a pointer to a C++
object that cannot be saved for reuse in a different session.
}
\seealso{
\code{\link{readImg}}, \code{\link{getFrame}}, \code{\link{imshow}}
}
\keyword{datasets}

