% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\docType{data}
\name{vpStream}
\alias{vpStream}
\title{Video stream object.}
\format{A vpStream object contains one field:
\itemize{
 \item{"dim": the dimensions in pixels (width x height) of the video stream.}
}}
\description{
A vpStream object is an R container for an OpenCV camera stream. vpStream
objects can be generated by the \code{\link{readStream}} function in the
\code{\link{videoplayR}} library.
}
\details{
A vpStream object is not a persistent object, but a pointer to a C++
object that cannot be saved for reuse in a different session.
}
\seealso{
\code{\link{readStream}}, \code{\link{nextFrame}}, \code{\link{release}}
}
\keyword{datasets}

