% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{writeImg}
\alias{writeImg}
\title{Save image object to file}
\usage{
writeImg(filename, image)
}
\arguments{
\item{filename}{A character string of the path to the image file.}

\item{image}{The vpImage object to save.}
}
\value{
This function does not return anythin.
}
\description{
This function write a vpImage object to an image file on the computer's drive.
This function can save image files in the following formats: 
\itemize{
 \item{Windows bitmaps - *.bmp, *.dib}
 \item{JPEG files - *.jpeg, *.jpg, *.jpe}
 \item{Portable Network Graphics - *.png}
 \item{Portable image format - *.pbm, *.pgm, *.ppm}
 \item{TIFF files - *.tiff, *.tif}
}
The image format is chosen based on the \code{filename} extension.
}
\examples{
filename <- system.file("sample_img/SampleVideo_1080x720_5mb.png", package = "videoplayR")
img <- readImg(filename)
writeImg("test_img.png", img)
}
\seealso{
\code{\link{readImg}}, \code{\link{vpImage}}
}

