% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io.R
\docType{data}
\name{vpVideo}
\alias{vpVideo}
\title{Video object.}
\format{A vpVideo object contains three fields:
\itemize{
 \item{"length": the number of frames in the video.}
 \item{"fps": the framerate of the video.}
 \item{"dim": the dimensions in pixels (width x height) of the video.}
}}
\description{
A vpVideo object is an R container for an OpenCV video. vpVideo objects can
be generated by the \code{\link{readVid}} function in the
\code{\link{videoplayR}} library.
}
\seealso{
\code{\link{readVid}}, \code{\link{getFrame}}
}
\keyword{datasets}

